refactor: moved template files and functions
