+++ data/template/kuboard.yaml
